# [expo-media-library][Android] fix Exif info loading