--- a/packages/expo-media-library/CHANGELOG.md
+++ b/packages/expo-media-library/CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### 🐛 Bug fixes
 
+- [Android] Fix `getAssetsAsync` loading performance, do not load full EXIF data for images by default ([#37957](https://github.com/expo/expo/pull/37957) by [@kosmydel](https://github.com/**kosmydel**))
 - [Android] Fix `deleteAssetsAsync` not working on android 11 or above ([#33211](https://github.com/expo/expo/pull/33211) by [@Zeeshan404](https://github.com/Zeeshan404))
 
 ## 17.1.6 — 2025-04-30