--- a/packages/expo-media-library/android/src/main/java/expo/modules/medialibrary/assets/AssetUtils.kt
+++ b/packages/expo-media-library/android/src/main/java/expo/modules/medialibrary/assets/AssetUtils.kt
@@ -110,7 +110,7 @@ fun putAssetsInfo(
     val localUri = "file://$path"
     val mediaType = cursor.getInt(mediaTypeIndex)
     var exifInterface: ExifInterface? = null
-    if (mediaType == MediaStore.Files.FileColumns.MEDIA_TYPE_IMAGE) {
+    if (resolveWithFullInfo && mediaType == MediaStore.Files.FileColumns.MEDIA_TYPE_IMAGE) {
       try {
         exifInterface = ExifInterface(path)
       } catch (e: IOException) {