--- a/packages/expo-media-library/android/src/main/java/expo/modules/medialibrary/assets/GetAssets.kt
+++ b/packages/expo-media-library/android/src/main/java/expo/modules/medialibrary/assets/GetAssets.kt
@@ -39,7 +39,7 @@ internal class GetAssets(
           assetsInfo,
           limit.toInt(),
           offset,
-          assetOptions.resolveWithFullInfo ?: false,
+          assetOptions.resolveWithFullInfo ?: false
         )
         val response = Bundle().apply {
           putParcelableArrayList("assets", assetsInfo)