--- a/packages/expo-media-library/android/src/test/java/expo/modules/medialibrary/assets/GetAssetsTest.kt
+++ b/packages/expo-media-library/android/src/test/java/expo/modules/medialibrary/assets/GetAssetsTest.kt
@@ -40,7 +40,8 @@ internal class GetAssetsTest {
       createdBefore = null,
       sortBy = emptyList(),
       mediaType = emptyList(),
-      album = null
+      album = null,
+      resolveWithFullInfo = false
     )
 
   @Before