--- a/packages/expo-media-library/build/MediaLibrary.js
+++ b/packages/expo-media-library/build/MediaLibrary.js
@@ -388,7 +388,7 @@ export async function getAssetsAsync(assetsOptions = {}) {
     if (!MediaLibrary.getAssetsAsync) {
         throw new UnavailabilityError('MediaLibrary', 'getAssetsAsync');
     }
-    const { first, after, album, sortBy, mediaType, createdAfter, createdBefore, mediaSubtypes } = assetsOptions;
+    const { first, after, album, sortBy, mediaType, createdAfter, createdBefore, mediaSubtypes, resolveWithFullInfo, } = assetsOptions;
     const options = {
         first: first == null ? 20 : first,
         after: getId(after),
@@ -398,6 +398,7 @@ export async function getAssetsAsync(assetsOptions = {}) {
         mediaSubtypes: arrayize(mediaSubtypes),
         createdAfter: dateToNumber(createdAfter),
         createdBefore: dateToNumber(createdBefore),
+        resolveWithFullInfo: resolveWithFullInfo ?? false,
     };
     if (first != null && typeof options.first !== 'number') {
         throw new Error('Option "first" must be a number!');