… docs/public/static/data/unversioned/expo-media-library.json
… packages/expo-media-library/build/MediaLibrary.d.ts.map
… packages/expo-media-library/build/MediaLibrary.js.map