# Why

<!--
Please describe the motivation for this PR, and link to relevant GitHub issues, forums posts, or feature requests.
-->
Loading a large number of images on Android was significantly slower compared to iOS. For example, loading ~2000 photos took 94 ms on an iPhone 16, but 9394 ms on a Pixel 9 Pro 5G.


# How

<!--
How did you build this feature or fix this bug and why?
-->
On Android we've integrated for all files and get the `ExifInterface` for each photo. It was later used in two places:
- `getAssetDimensionsFromCursor` for fixing images orientation,
- and in `getExifFullInfo` (but this is disabled without `resolveWithFullInfo` flag).

On Android, we were previously loading ExifInterface for every photo, regardless of whether full EXIF data was needed. 

This was used in two places:
- In [`getAssetDimensionsFromCursor`](https://github.com/expo/expo/blob/66d2a7cbc7b046a546be0456f4f384b692611fe7/packages/expo-media-library/android/src/main/java/expo/modules/medialibrary/assets/AssetUtils.kt#L121) to correct image orientation.
- In [`getExifFullInfo`](https://github.com/expo/expo/blob/66d2a7cbc7b046a546be0456f4f384b692611fe7/packages/expo-media-library/android/src/main/java/expo/modules/medialibrary/assets/AssetUtils.kt#L135) (only when `resolveWithFullInfo` is enabled).

This PR avoids loading full EXIF info unless the `resolveWithFullInfo` flag is set. As a result, some images may no longer be oriented correctly on certain Android devices when this flag is not enabled. However, the performance improvement is significant — up to a 10x speed increase.

| Device | Without fix (ms) | With fix (ms) | Speedup |
|--------|--------|--------|--------|
| Pixel 9 Pro 5G | 9394 | 524 | 17.93 |
| Pixel 4a | 16591 | 1588 | 10.45 |

> [!NOTE]  
> Alternative solution: We can add a separate prop for that.

# Test Plan

<!--
Please describe how you tested this change and how a reviewer could reproduce your test, especially if this PR does not include automated tests! If possible, please also provide terminal output and/or screenshots demonstrating your test/reproduction.
-->

1. Fetch all assets on the Android device.
2. Verify its performance.
3. If there are problems with image orientation, use `resolveWithFullInfo` prop to enable EXIF fix.

| `resolveWithFullInfo = false` | `= true` |
|--------|--------|
| <img src="https://github.com/user-attachments/assets/55c7ca0a-a678-4350-a750-5ab8040a056b" width="350"> | <img src="https://github.com/user-attachments/assets/222d06b2-9cf2-4a7f-9cc7-63c817cf059a" width="350"> |

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [x] I added a `changelog.md` entry and rebuilt the package sources according to [this short guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [x] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).
- [x] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)